requirements.txt
pandas>=2.0
requests>=2.31
boto3>=1.34
python-dotenv>=1.0   #to load .env locally